amélioration du code

+++ frontend/js/confirmation.js
@@ -1,20 +1,22 @@
 //Récupération localstorage
 let commande =  Object.keys(localStorage);
 let infoCommand = JSON.parse(localStorage.getItem(commande))
+if (commande == 0){
+    let commandeNull = document.getElementById('commandeNull')
+    commandeNull.className = "commandeNull"
+    commandeNull.innerHTML = "Vous n\' avez pas de commande !"
+}else {
+    let thanks = document.getElementById('h2_thanks')
+    thanks.textContent = `Confirmation de votre commande ${infoCommand.nameConfirmation}, merci pour votre achat !`;
 
-let divConfirmation = document.getElementById('div_confirmation') 
+    let info = document.getElementById('info')
+    info.textContent = "Récapitulatif de votre commande :";
 
-let thanks = document.getElementById('h2_thanks')
-thanks.textContent = `Confirmation de votre commande ${infoCommand.nameConfirmation}, merci pour votre achat!`;
+    let id = document.getElementById('id_commande')
+    id.innerHTML = `Numéro de commande : ${infoCommand.orderId}`;
 
-let info = document.getElementById('info')
-info.textContent = "Récapitulatif de votre commande :";
+    let price = document.getElementById('priceConfirmation')
+    price.innerHTML = `Prix total de la commande : ${infoCommand.totalPriceBasket} €`;
 
-let id = document.getElementById('id_commande')
-id.innerHTML = `Numéro de commande : ${infoCommand.orderId}`;
-
-let price = document.getElementById('priceConfirmation')
-price.innerHTML = `Prix total de la commande : ${infoCommand.totalPriceBasket} €`;
-
-localStorage.clear();
-
+    localStorage.clear();
+}
